Started Basic Setup
Added the "feed" application which will be a feed of Events for volunteers.

Migrated to SQLite

Started on Utility models including Profile (a subclass of User that will later contain important information), Event (a class that is instantiated for new Events. This would be like: "Marathon") and Slot (a class that is instantiated for time slots. This would be like: "Water Bottle Stand, 3pm-5pm")