Added sign up system
Can be found at /accounts/signup or via a ling from the login page

Usernames and emails must be unique to register. The account is added to
the database instantly but can only be logged into once the user has
confirmed their email. The password fields must match for the form to
return valid.

Added the email to settings.py to send info to clients including
registration confirmation. Ask Armaan or Joe for the password if you
need it for debugging.